% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{check_discrepancies}
\alias{check_discrepancies}
\title{Checks for discrepancies between S1 and S2 measurements in the DataFrame and prints them to the console.
This function is not included in the big pre-processing function, as it is more intended to
perform a quality check on your data and not to automatically inform the processing of the data.}
\usage{
check_discrepancies(df, threshold = 0.05, individual = FALSE)
}
\arguments{
\item{df}{DataFrame containing wric data with columns for S1 and S2 measurements.}

\item{threshold}{Numeric threshold percentage for mean relative delta discrepancies (default 0.05).}

\item{individual}{Logical, if TRUE checks and reports individual row discrepancies beyond the threshold (default FALSE).}
}
\value{
None. Prints discrepancies to the console.
}
\description{
Checks for discrepancies between S1 and S2 measurements in the DataFrame and prints them to the console.
This function is not included in the big pre-processing function, as it is more intended to
perform a quality check on your data and not to automatically inform the processing of the data.
}
