% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{combine_measurements}
\alias{combine_measurements}
\title{Combines S1 and S2 measurements in the DataFrame using the specified method.}
\usage{
combine_measurements(df, method = "mean")
}
\arguments{
\item{df}{DataFrame containing wric data with S1 and S2 measurement columns.}

\item{method}{String specifying the method to combine measurements ("mean", "median", "s1", "s2", "min", "max").}
}
\value{
A DataFrame with combined measurements.
}
\description{
Combines S1 and S2 measurements in the DataFrame using the specified method.
}
