% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{cut_rows}
\alias{cut_rows}
\title{Filters rows in a DataFrame based on an optional start and end datetime range.}
\usage{
cut_rows(df, start = NULL, end = NULL)
}
\arguments{
\item{df}{data.frame
DataFrame with a "datetime" column to filter.}

\item{start}{character or POSIXct or NULL, optional;
Start datetime; rows before this will be removed. If NULL, uses the earliest datetime in the DataFrame.}

\item{end}{character or POSIXct or NULL, optional
End datetime; rows after this will be removed. If NULL, uses the latest datetime in the DataFrame.}
}
\value{
data.frame
DataFrame with rows between the specified start and end dates, or the full DataFrame if both are NULL.
}
\description{
Filters rows in a DataFrame based on an optional start and end datetime range.
}
\examples{
df <- data.frame(
  datetime = as.POSIXct(c(
    "2023-11-13 11:40:00",
    "2023-11-13 11:45:00",
    "2023-11-13 11:50:00"
  ))
)

# Filter rows from 11:45 onward
cut_rows(df, start = "2023-11-13 11:45:00")

# Filter rows between 11:40 and 11:45
cut_rows(df, start = "2023-11-13 11:40:00", end = "2023-11-13 11:45:00")

# No filtering (both NULL)
cut_rows(df)
}
