% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{detect_start_end}
\alias{detect_start_end}
\title{Automatically detect enter and exit from the chamber based on the notefile.
Returns the start and end times for two participants.}
\usage{
detect_start_end(notes_path)
}
\arguments{
\item{notes_path}{string - path to the note file}
}
\value{
list - A list of two elements ("1" and "2"), each containing a tuple (start, end) time.
Returns NA if not possible to find start or end time.
}
\description{
Automatically detect enter and exit from the chamber based on the notefile.
Returns the start and end times for two participants.
}
\keyword{chamber}
\keyword{detection}
\keyword{entry}
\keyword{exit}
