% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{export_file_from_redcap}
\alias{export_file_from_redcap}
\title{Exports a file from REDCap based on the specified record ID and field name.}
\usage{
export_file_from_redcap(record_id, fieldname, path = NULL, api_url, api_token)
}
\arguments{
\item{record_id}{String containing the unique identifier for the record in REDCap.}

\item{fieldname}{Field name from which to export the file.}

\item{path}{File path where the exported file will be saved.}

\item{api_url}{String, URL to the REDCap API, should be specified in your personal config.R file}

\item{api_token}{String, personal token for the REDCap API, should be specified in your personal config.R file}
}
\value{
filepath String, the filepath that the file was just saved to.
}
\description{
If you do not specify a path, the data will be downloaded to a temporary folder
which is deleted when your R session ends.
}
\examples{
\dontshow{if (file.exists(path.expand("~/.config.R"))) withAutoprint(\{ # examplesIf}
source(path.expand("~/.config.R"))
export_file_from_redcap(
  record_id = "1", fieldname = "wric_data",
  api_url = api_url, api_token = api_token
)
\dontshow{\}) # examplesIf}
}
