% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{extract_meta_data}
\alias{extract_meta_data}
\title{Extracts metadata for two subjects from text lines and optionally saves it as CSV files.}
\usage{
extract_meta_data(lines, code, manual, save_csv, path_to_save)
}
\arguments{
\item{lines}{List of strings containing the WRIC metadata.}

\item{code}{Method for generating subject IDs ("id", "id+comment", or "manual").}

\item{manual}{Custom codes for Room 1 and Room 2 subjects if \code{code} is "manual".}

\item{save_csv}{Logical, whether to save extracted metadata to CSV files.}

\item{path_to_save}{Directory path for saving CSV files, NULL uses the current directory.}
}
\value{
A list containing the Room 1 code, Room 2 code, and DataFrames for R1_metadata and R2_metadata.
}
\description{
Extracts metadata for two subjects from text lines and optionally saves it as CSV files.
}
