% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{extract_metadata_new}
\alias{extract_metadata_new}
\title{Extracts metadata (software v2) for a single subject from text lines and optionally saves it as a CSV file.}
\usage{
extract_metadata_new(
  lines,
  code,
  manual = NULL,
  save_csv = FALSE,
  path_to_save = NULL
)
}
\arguments{
\item{lines}{List of strings containing the wric metadata.}

\item{manual}{Custom code for the subject if \code{code} is "manual".}

\item{save_csv}{Logical, whether to save extracted metadata to a CSV file.}

\item{path_to_save}{Directory path for saving the CSV file, NULL uses the current directory.}
}
\value{
A list containing:
\describe{
\item{code}{The generated subject code.}
\item{metadata}{A data.frame containing the extracted metadata.}
}
}
\description{
Extracts metadata (software v2) for a single subject from text lines and optionally saves it as a CSV file.
}
\examples{
lines <- c(
  "2.0.0.15\tOmnical software by Maastricht Instruments B.V.",
  "file identifier is C:\\\\Omnical\\\\Results Room 1\\\\1 min\\\\Results.txt",
  "Calibration values\t18.0\t0.795",
  "",
  "Subject ID\tStudy ID\tMeasurement ID\tResearcher ID\tComments",
  "ZeroTest\t\t\tNZ\tZero Test 14.01.2026"
)

# Extract metadata and generate code from subject ID only
extract_metadata_new(lines, code = "id", manual = NULL, save_csv = FALSE)

# Extract metadata using ID + comment as code
extract_metadata_new(lines, code = "id+comment", manual = NULL, save_csv = FALSE)

# Extract metadata using a manual code
extract_metadata_new(lines, code = "manual", manual = "custom_code", save_csv = FALSE)
}
