% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{open_file}
\alias{open_file}
\title{Opens a WRIC .txt file and reads its contents.}
\usage{
open_file(filepath)
}
\arguments{
\item{filepath}{Path to the WRIC .txt file.}
}
\value{
A list of strings representing the lines of the file.
}
\description{
Opens a WRIC .txt file and reads its contents.
}
\note{
Raises an error if the file is not a valid WRIC data file.
}
