% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{preprocess_wric_files}
\alias{preprocess_wric_files}
\title{Preprocesses multiple wric_files by RedCAP record ID, extracting metadata, creating DataFrames, and optionally saving results.}
\usage{
preprocess_wric_files(
  csv_file,
  fieldname,
  code = "id",
  manual = NULL,
  save_csv = TRUE,
  path_to_save = NULL,
  combine = TRUE,
  method = "mean",
  start = NULL,
  end = NULL
)
}
\arguments{
\item{csv_file}{Path to the CSV file containing record IDs.}

\item{fieldname}{The field name for exporting wric data from RedCAP.}

\item{code}{Method for generating subject IDs ("id", "id+comment", or "manual").}

\item{manual}{Custom codes for subjects in Room 1 and Room 2 if \code{code} is "manual".}

\item{save_csv}{Logical, whether to save extracted metadata and data to CSV files.}

\item{path_to_save}{Directory path for saving CSV files, NULL uses the current directory.}

\item{combine}{Logical, whether to combine S1 and S2 measurements.}

\item{method}{Method for combining measurements ("mean", "median", "s1", "s2", "min", "max").}

\item{start}{character or POSIXct or NULL, rows before this will be removed, if NULL takes first row e.g "2023-11-13 11:43:00"}

\item{end}{character or POSIXct or NULL, rows after this will be removed, if NULL takes last row e.g "2023-11-13 11:43:00"}
}
\value{
A list where each key is a record ID and each value is a list with: (r1_metadata, r2_metadata, df_room1, df_room2).
}
\description{
Preprocesses multiple wric_files by RedCAP record ID, extracting metadata, creating DataFrames, and optionally saving results.
}
