% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{upload_file_to_redcap}
\alias{upload_file_to_redcap}
\title{Uploads a file to REDCap for a specified record ID and field name.}
\usage{
upload_file_to_redcap(filepath, record_id, fieldname, api_url, api_token)
}
\arguments{
\item{filepath}{Path to the file to be uploaded.}

\item{record_id}{String containing the unique identifier for the record in REDCap.}

\item{fieldname}{Field name to which the file will be uploaded.}

\item{api_url}{String, URL to the REDCap API, should be specified in your personal config.R file}

\item{api_token}{String, personal token for the REDCap API, should be specified in your personal config.R file}
}
\value{
None. Prints the HTTP status code of the request.
}
\description{
Uploads a file to REDCap for a specified record ID and field name.
}
