% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrictools-package.R
\docType{package}
\name{wrictools-package}
\alias{wrictools}
\alias{wrictools-package}
\title{wrictools: Analyze Whole Room Indirect Calorimetery (WRIC) Data}
\description{
Description: Provides functions, tutorials, and examples to preprocess, analyze, and visualize data from whole room indirect calorimeters (WRIC) by Maastricht Instruments, using the 'OmniCal' software. Some functions may also work with WRICs from other manufacturers, though full functionality has only been validated for Maastricht Instruments devices.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/NinaZiegenbein/wrictools}
  \item \url{https://ninaziegenbein.github.io/wrictools/}
  \item Report bugs at \url{https://github.com/NinaZiegenbein/wrictools/issues}
}

}
\author{
\strong{Maintainer}: Nina Ziegenbein \email{nina.ziegenbein@web.de} (\href{https://orcid.org/0009-0005-4913-8910}{ORCID}) [copyright holder]

}
\keyword{internal}
