% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{check_code}
\alias{check_code}
\title{Check the subject ID code and return corresponding Room 1 and Room 2 codes.}
\usage{
check_code(code, manual, r1_metadata, r2_metadata)
}
\arguments{
\item{code}{Method for generating subject IDs ("id", "id+comment", or "manual").}

\item{manual}{A list of custom codes for Room 1 and Room 2, required if \code{code} is "manual".}

\item{r1_metadata}{DataFrame for metadata of Room 1, containing "Subject ID" and "Comments".}

\item{r2_metadata}{DataFrame for metadata of Room 2, containing "Subject ID" and "Comments".}
}
\value{
A list containing the codes for Room 1 and Room 2.
}
\description{
Check the subject ID code and return corresponding Room 1 and Room 2 codes.
}
