% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{create_wric_df}
\alias{create_wric_df}
\title{Creates DataFrames for wric data from a file and optionally saves them as CSV files.}
\usage{
create_wric_df(
  filepath,
  lines,
  save_csv,
  code_1,
  code_2,
  path_to_save,
  start,
  end,
  notefilepath
)
}
\arguments{
\item{filepath}{Path to the wric .txt file.}

\item{lines}{List of strings read from the file to locate the data start.}

\item{save_csv}{Logical, whether to save DataFrames as CSV files.}

\item{code_1}{String representing the codes for Room 1.}

\item{code_2}{String representing the codes for Room 2.}

\item{path_to_save}{Directory path for saving CSV files, NULL uses the current directory.}

\item{start}{character or POSIXct or NULL, rows before this will be removed, if NULL takes first row e.g "2023-11-13 11:43:00"}

\item{end}{character or POSIXct or NULL, rows after this will be removed, if NULL takes last row e.g "2023-11-13 11:43:00"}

\item{notefilepath}{String, The path to the notefile}
}
\value{
A list containing DataFrames for Room 1 and Room 2 measurements.
}
\description{
Creates DataFrames for wric data from a file and optionally saves them as CSV files.
}
\note{
Raises an error if Date or Time columns are inconsistent across rows.
}
