% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{cut_rows}
\alias{cut_rows}
\title{Filters rows in a DataFrame based on an optional start and end datetime range.}
\usage{
cut_rows(df, start = NULL, end = NULL)
}
\arguments{
\item{df}{data.frame
DataFrame with a "datetime" column to filter.}

\item{start}{character or POSIXct or NULL, optional;
Start datetime; rows before this will be removed. If NULL, uses the earliest datetime in the DataFrame.}

\item{end}{character or POSIXct or NULL, optional
End datetime; rows after this will be removed. If NULL, uses the latest datetime in the DataFrame.}
}
\value{
data.frame
DataFrame with rows between the specified start and end dates, or the full DataFrame if both are NULL.
}
\description{
Filters rows in a DataFrame based on an optional start and end datetime range.
}
