% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{export_file_from_redcap}
\alias{export_file_from_redcap}
\title{Exports a file from REDCap based on the specified record ID and field name.}
\usage{
export_file_from_redcap(
  record_id,
  fieldname,
  path = "./tmp/export.raw.txt",
  api_url,
  api_token
)
}
\arguments{
\item{record_id}{String containing the unique identifier for the record in REDCap.}

\item{fieldname}{Field name from which to export the file.}

\item{path}{File path where the exported file will be saved.}

\item{api_url}{String, URL to the REDCap API, should be specified in your personal config.R file}

\item{api_token}{String, personal token for the REDCap API, should be specified in your personal config.R file}
}
\value{
None. The file is saved to the specified path.
}
\description{
Exports a file from REDCap based on the specified record ID and field name.
}
