% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{extract_meta_data}
\alias{extract_meta_data}
\title{Extracts metadata for two subjects from text lines and optionally saves it as CSV files.}
\usage{
extract_meta_data(lines, code, manual, save_csv, path_to_save)
}
\arguments{
\item{lines}{List of strings containing the wric metadata.}

\item{code}{Method for generating subject IDs ("id", "id+comment", or "manual").}

\item{manual}{Custom codes for Room 1 and Room 2 subjects if \code{code} is "manual".}

\item{save_csv}{Logical, whether to save extracted metadata to CSV files.}

\item{path_to_save}{Directory path for saving CSV files, NULL uses the current directory.}
}
\value{
A list containing the Room 1 code, Room 2 code, and DataFrames for r1_metadata and r2_metadata.
}
\description{
Extracts metadata for two subjects from text lines and optionally saves it as CSV files.
}
\examples{
lines <- c(
  "OmniCal software by ing.P.F.M.Schoffelen, Dept. of Human Biology, Maastricht University",
  "file identifier is C:\\\\MI_Room_Calorimeter\\\\Results_online\\\\1_minute\\\\Results.txt",
  "",
  "Room 1\tProject\tSubject ID\tExperiment performed by\tComments",
  "\tPROJECT\tXXXX\tJANE DOE\t",
  "Room 2\tProject\tSubject ID\tExperiment performed by\tComments",
  "\tPROJECT\tYYYY\tJOHN DOE\t"
)

extract_meta_data(lines, code = "id", manual = NULL, save_csv = FALSE, path_to_save = NULL)
}
