% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{extract_note_info}
\alias{extract_note_info}
\title{Extracts and processes note information from a specified notes file, categorizing events
based on predefined keywords, and updates two DataFrames with protocol information for
different participants.}
\usage{
extract_note_info(notes_path, df_room1, df_room2, keywords_dict = NULL)
}
\arguments{
\item{notes_path}{string - The file path to the notes file containing event data.}

\item{df_room1}{DataFrame - DataFrame for participant 1, to be updated with protocol info.}

\item{df_room2}{DataFrame - DataFrame for participant 2, to be updated with protocol info.}

\item{keywords_dict}{nested list - used to identify keywords to extract protocol values}
}
\value{
list - A list containing two updated DataFrames:
- \code{df_room1}: Updated DataFrame for participant 1 with protocol data.
- \code{df_room2}: Updated DataFrame for participant 2 with protocol data.
}
\description{
Extracts and processes note information from a specified notes file, categorizing events
based on predefined keywords, and updates two DataFrames with protocol information for
different participants.
}
\note{
\itemize{
\item The 'Comment' field should start with '1' or '2' to indicate the participant,
or it may be empty to indicate both.
\item The \code{keywords_dict} can be modified to fit specific study protocols,
with multi-group checks for keyword matching.
}
}
