% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{open_file}
\alias{open_file}
\title{Opens a WRIC .txt file and reads its contents, identifying the software version.}
\usage{
open_file(filepath)
}
\arguments{
\item{filepath}{description filepath}
}
\value{
A list with two elements:
\itemize{
\item lines: the lines of the file
\item v1: TRUE if version 1, FALSE if version 2
}
}
\description{
Opens a WRIC .txt file and reads its contents, identifying the software version.
}
\note{
Raises an error if the file is not a valid wric data file.
}
\keyword{internal}
