% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{open_file}
\alias{open_file}
\title{Opens a wric .txt file and reads its contents.}
\usage{
open_file(filepath)
}
\arguments{
\item{filepath}{description filepath}
}
\value{
A list of strings representing the lines of the file.
}
\description{
Opens a wric .txt file and reads its contents.
}
\note{
Raises an error if the file is not a valid wric data file.
}
\keyword{internal}
