% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{preprocess_wric_file}
\alias{preprocess_wric_file}
\title{Preprocesses a wric data file, extracting metadata, creating DataFrames, and optionally saving results.}
\usage{
preprocess_wric_file(
  filepath,
  code = "id",
  manual = NULL,
  save_csv = TRUE,
  path_to_save = NULL,
  combine = TRUE,
  method = "mean",
  start = NULL,
  end = NULL,
  notefilepath = NULL,
  keywords_dict = NULL
)
}
\arguments{
\item{filepath}{Path to the wric .txt file.}

\item{code}{Method for generating subject IDs ("id", "id+comment", or "manual").}

\item{manual}{Custom codes for subjects in Room 1 and Room 2 if \code{code} is "manual".}

\item{save_csv}{Logical, whether to save extracted metadata and data to CSV files.}

\item{path_to_save}{Directory path for saving CSV files, NULL uses the current directory.}

\item{combine}{Logical, whether to combine S1 and S2 measurements.}

\item{method}{Method for combining measurements ("mean", "median", "s1", "s2", "min", "max").}

\item{start}{character or POSIXct or NULL, rows before this will be removed, if NULL takes first row e.g "2023-11-13 11:43:00"}

\item{end}{character or POSIXct or NULL, rows after this will be removed, if NULL takes last row e.g "2023-11-13 11:43:00"}

\item{notefilepath}{String, Directory path of the corresponding note file (.txt)}

\item{keywords_dict}{Nested List, used to extract protocol values from note file}
}
\value{
A list containing the metadata and DataFrames for Room 1 and Room 2.
}
\description{
Preprocesses a wric data file, extracting metadata, creating DataFrames, and optionally saving results.
}
